apps/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.30.0
python-dotenv>=1.0.0
stripe>=11.0.0
httpx>=0.28.0
pydantic>=2.10.0
